requirements.txt
torch
sentence-transformers
faiss-cpu
langchain
fastapi
uvicorn